pyproject.toml
[tool.poetry]
include = ["MANIFEST.skao.int"]
name = "ska-integration-test-harness"
version = "0.1.0"
description = "For now, a test harness for TMC-CSP integration tests. In future, a generic test harness integration testing an arbitrary combination of production or emulated SKA subsystems."
authors = ["Emanuele Lena <emanuele.lena@designcoaching.net>"]
license = "BSD-3-Clause"
readme = "README.md"
packages = [
    { include = "ska_integration_test_harness", from = "src" }
]


[[tool.poetry.source]]
name = 'skao'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[[tool.poetry.source]]
name = "PyPI-public"
url = 'https://pypi.org/simple'


[tool.poetry.dependencies]
python = "^3.10"
assertpy = "^1.1"
pytango = "^9.5.1"
ska-control-model = "^1.0.0"
ska-tango-testing = "^0.7.1"
ska-ser-logging = "^0.4.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.4.2"
flake8 = "^7.1.0"
pylint = "^3.2.6"
pylint-junit = "^0.3.4"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
build = "^1.2.1"


[tool.poetry.group.docs.dependencies]
sphinx = "^7.4.7"
myst-parser = "^3.0.1"
ska-ser-sphinx-theme = "^0.2.0"
sphinx-autodoc-typehints = "^2.2.3"
typing-extensions = "^4.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
